Change captcha, fixed admin menu behaviour
As I noticed the previous CaptCha was not fully working in all cases, I decided to try and merge the newer CaptCha from 'grohsfabian/minecraft-servers-list-lite'. It ended up being a good decision, as it works.
I assume this is OK with you, grohsfabian, as the code is still from your similar project.

Also fixed wrong behaviour of admin dropdown menu. (Was kinda my previous commits' fault. Sorry about that (^^'))